+++ CHANGELOG.md
@@ -6,12 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Honeypot detection
+
 ### Added
 
 - Hostname can be used as a username or password dynamically ($TARGET) (#28)
 - Output log format are now customizable (#30)
 
 ### Fixed
+
 - Added missing wrapper for FD_SET in static build
 - ssh_get_error shouldnt be used after ssh_free
 - Buffer overflow in banner grabber